+++ src/index.ts
@@ -1,11 +1,5 @@
 export type { InferType } from './InferType';
 export { parse } from './parse';
 export { parseOrFail } from './parseOrFail';
-export { number } from './asserts/number';
-export { string } from './asserts/string';
-export { boolean } from './asserts/boolean';
-export { array } from './asserts/array';
-export { object } from './asserts/object';
-export { oneOfTypes } from './asserts/mix';
 export { ValidationError, BuildSchemaError, guardException } from './exceptions';
 export { setLocale, setToDefaultLocale, clearLocales } from './translationMap';